+++ src/org/sunflow/core/primitive/SphereFlake.java
@@ -25,7 +25,7 @@ public class SphereFlake implements PrimitiveList {
     static {
         // geometric series table, to compute bounding radius quickly
         for (int i = 0, r = 3; i < boundingRadiusOffset.length; i++, r *= 3)
-            boundingRadiusOffset[i] = (float) (r - 3.0f) / r;
+            boundingRadiusOffset[i] = (r - 3.0f) / r;
         // lower ring
         double a = 0, daL = 2 * Math.PI / 6, daU = 2 * Math.PI / 3;
         for (int i = 0; i < 6; i++) {
@@ -134,9 +134,9 @@ public class SphereFlake implements PrimitiveList {
                 if (t1 >= r.getMax() || t2 <= r.getMin())
                     return;
                 if (t1 > r.getMin())
-                    r.setMax((float) t1);
+                    r.setMax(t1);
                 else
-                    r.setMax((float) t2);
+                    r.setMax(t2);
                 state.setIntersection(0, cx, cy, cz);
             }
         } else {
@@ -166,9 +166,9 @@ public class SphereFlake implements PrimitiveList {
                         // no hit
                     } else {
                         if (t1 > r.getMin())
-                            r.setMax((float) t1);
+                            r.setMax(t1);
                         else
-                            r.setMax((float) t2);
+                            r.setMax(t2);
                         state.setIntersection(0, cx, cy, cz);
                     }
                 }
